rapifuzz-backend/requirements.txt
APScheduler==3.10.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
Django==4.2.6
django-appconf==1.0.5
django-cors-headers==3.14.0
django-db-connection-pool==1.2.2
django-filter==22.1
django-oauth-toolkit==2.2.0
django-request==1.6.2
django-rest-framework==0.1.0
django-staticfiles==1.2.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dict-digger==0.2.1
drf-standardized-errors==0.12.5
drf-yasg==1.21.5
flatten-dict==0.4.2
gunicorn==20.1.0
iteration-utilities==0.11.0
json5==0.9.11
licensing==0.39
matplotlib
matplotlib-inline==0.1.6
mitmproxy
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
pandas-ods-reader==0.1.4
pdfkit==1.0.0
pycryptodome==3.17
PyJWT==2.6.0
pypdf2==3.0.1
PySocks==1.7.1
PyYAML==6.0
requests==2.28.2
sockets==1.0.0
urllib3==1.26.15
validators==0.20.0
weasyprint==58.1
Werkzeug==2.2.3
whitenoise==6.4.0
XlsxWriter==3.0.9
xmltodict==0.13.0
zeep==4.2.1
toml
graphql-core==3.2.3
GraphQL-core-next==1.1.1
